@@ -1,37 +1,55 @@
 {
-  "name": "melon",
-  "version": "1.0.0",
-  "description": "Melon is a nuxt.js starter for creative development, using Three, Pixi, SmoothScroll and i18n, made by Antoine Abbou",
-  "author": "Antoine Abbou <abbouantoine@gmail.com>",
+  "name": "antoineabbou",
+  "version": "2.0.0",
+  "description": "Portfolio 2019 - Made with Nuxt.JS",
+  "author": "Antoine Abbou",
   "private": true,
   "scripts": {
-    "dev": "cross-env HOST=0.0.0.0 PORT=3000 nuxt",
+    "dev": "cross-env HOST=0.0.0.0 nuxt",
     "build": "nuxt build",
-    "start": "nuxt start",
     "generate": "nuxt generate",
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
-    "precommit": "npm run lint"
+    "precommit": "npm run lint",
+    "start": "nuxt start"
   },
   "dependencies": {
-    "gsap": "^2.0.1",
-    "nuxt-edge": "^2.0.0-25494410.e993616",
-    "pixi.js": "^4.7.0",
-    "smooth-scrolling": "^2.3.11",
-    "three": "^0.95.0",
-    "vue-i18n": "^8.0.0"
+    "@brunoimbrizi/controlkit": "^0.1.97",
+    "@nuxtjs/axios": "^5.3.6",
+    "browser-detect": "^0.2.28",
+    "cross-env": "^5.2.0",
+    "domready": "^1.0.8",
+    "gsap": "^2.1.3",
+    "lodash.throttle": "^4.1.1",
+    "mobile-detect": "^1.4.3",
+    "nuxt": "^2.0.0",
+    "nuxt-payload-extractor": "^0.0.9",
+    "pixi.js": "^5.1.0",
+    "slugify": "^1.3.4",
+    "smooth-scrolling": "^2.3.12",
+    "three": "^0.106.2",
+    "uuid": "^3.3.2",
+    "vue-i18n": "^8.12.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^8.2.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "cross-env": "^5.1.3",
-    "eslint": "^4.15.0",
-    "eslint-friendly-formatter": "^3.0.0",
-    "eslint-loader": "^1.7.1",
-    "eslint-plugin-vue": "^4.0.0",
+    "@nuxtjs/eslint-config": "^0.0.1",
+    "@nuxtjs/eslint-module": "^0.0.1",
+    "babel-eslint": "^10.0.1",
+    "eslint": "^5.15.1",
+    "eslint-config-standard": ">=12.0.0",
+    "eslint-plugin-import": ">=2.16.0",
+    "eslint-plugin-jest": ">=22.3.0",
+    "eslint-plugin-node": ">=8.0.1",
+    "eslint-plugin-nuxt": ">=0.4.2",
+    "eslint-plugin-promise": ">=4.0.1",
+    "eslint-plugin-standard": ">=4.0.0",
+    "eslint-plugin-vue": "^5.2.2",
+    "file-loader": "^1.1.11",
+    "glslify": "^6.2.1",
+    "imports-loader": "^0.8.0",
+    "nodemon": "^1.18.9",
+    "pug": "^2.0.3",
+    "pug-plain-loader": "^1.0.0",
+    "raw-loader": "^0.5.1",
     "rupture": "^0.7.1",
     "stylus": "^0.54.5",
     "stylus-loader": "^3.0.2"
