@@ -8,14 +8,13 @@ module.exports = {
     parser: 'babel-eslint'
   },
   extends: [
-    // https://github.com/vuejs/eslint-plugin-vue#priority-a-essential-error-prevention
-    // consider switching to `plugin:vue/strongly-recommended` or `plugin:vue/recommended` for stricter rules.
-    'plugin:vue/essential'
+    '@nuxtjs',
+    'plugin:nuxt/recommended'
   ],
-  // required to lint *.vue files
   plugins: [
-    'vue'
   ],
   // add your custom rules here
-  rules: {}
+  rules: {
+    "no-console": "off"
+  }
 }
