@@ -1,5 +1,5 @@
 import { EventEmitter } from 'events'
 
-let emitter = new EventEmitter()
+const emitter = new EventEmitter()
 emitter.setMaxListeners(50)
 export default emitter
