@@ -4,12 +4,12 @@ export default function ({ isHMR, app, store, route, params, error, redirect }) 
   if (isHMR) return
   // Get locale from params
   const locale = params.lang || defaultLocale
-  if (store.state.locales.indexOf(locale) === -1) {
+  if (store.getters['locale/all'].indexOf(locale) === -1) {
     return error({ message: 'This page could not be found.', statusCode: 404 })
   }
   // Set locale
-  store.commit('SET_LANG', locale)
-  app.i18n.locale = store.state.locale
+  store.commit('locale/SET_LOCALE', locale)
+  app.i18n.locale = store.getters['locale/current']
   // If route is /<defaultLocale>/... -> redirect to /...
   if (locale === defaultLocale && route.fullPath.indexOf('/' + defaultLocale) === 0) {
     const toReplace = '^/' + defaultLocale + (route.fullPath.indexOf('/' + defaultLocale + '/') === 0 ? '/' : '')
