@@ -1,4 +1,4 @@
 <script>
-  import About from '~/pages/_lang/about'
-  export default About
+import About from '~/pages/_lang/about'
+export default About
 </script>
