@@ -1,4 +1,4 @@
 <script>
-  import Index from '~/pages/_lang/index'
-  export default Index
+import Index from '~/pages/_lang/index'
+export default Index
 </script>
